% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R
\name{as.fd.list}
\alias{as.fd.list}
\alias{as.list.fd}
\alias{as.fd.fd}
\alias{as.fd.dd}
\title{Coerce to / from an \code{fd} object}
\usage{
\method{as.fd}{list}(x, merge = FALSE, ...)

\method{as.list}{fd}(x, ...)

\method{as.fd}{fd}(x, ...)

\method{as.fd}{dd}(x, ...)
}
\arguments{
\item{x}{Object to coerce.}

\item{merge}{Pass-through to \code{\link[fda:as.fd]{fda::as.fd()}}.}

\item{...}{Further arguments forwarded to the underlying method.}
}
\description{
Coerce to / from an \code{fd} object
}
