% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{as.list.dd}
\alias{as.list.dd}
\title{Coerce a \code{dd} object to a list}
\usage{
\method{as.list}{dd}(x, ...)
}
\arguments{
\item{x}{A \code{dd} object.}

\item{...}{Currently ignored.}
}
\value{
A \code{ddl} list (also of class \code{fdl} / \code{list}).
}
\description{
Splits a multi-realization \code{dd} object into a list (one element per
column of \code{x$coefs}).
}
