% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{as_dd}
\alias{as_dd}
\alias{as_dd.list}
\alias{as_dd.xts}
\alias{as_dd.dd}
\alias{as_dd.fd}
\alias{as_dd.numeric}
\title{Coerce to a distributional-data object}
\usage{
as_dd(x, ...)

\method{as_dd}{list}(x, full_sample = unlist(x), mc.cores = NULL, ...)

\method{as_dd}{xts}(x, ...)

\method{as_dd}{dd}(x, ...)

\method{as_dd}{fd}(x, ...)

\method{as_dd}{numeric}(x, ...)
}
\arguments{
\item{x}{Object to coerce: a numeric sample, an \code{fd}, \code{dd}, list, or
\code{xts} time series.}

\item{...}{Further arguments forwarded to \code{\link[=dd]{dd()}}.}

\item{full_sample}{Pooled reference sample used to set common breakpoints
across list elements. Defaults to \code{unlist(x)}.}

\item{mc.cores}{Optional integer; if non-\code{NULL}, list elements are
processed in parallel via \code{\link[parallel:mclapply]{parallel::mclapply()}}.}
}
\value{
A \code{dd} object, or a list of them with class \code{c("ddl", "fdl", "list")} for list / xts input.
}
\description{
Coerce to a distributional-data object
}
