% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{boot.diff}
\alias{boot.diff}
\title{Bootstrap Difference helper function}
\usage{
boot.diff(dat, g)
}
\arguments{
\item{dat}{a data frame containing the predictor and outcome}

\item{g}{a vector of indices}

\item{...}{other arguments not used by this method.}
}
\value{
Invisibly returns the difference between the skewness and kurtosis of two vectors.
}
\description{
Calculates the bootstrap difference between two vectors.
}
