% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R, R/fd.R
\name{c.dd}
\alias{c.dd}
\alias{c.ddl}
\alias{c.fdl}
\title{Concatenate distributional / functional objects}
\usage{
\method{c}{dd}(...)

\method{c}{ddl}(...)

\method{c}{fdl}(...)
}
\arguments{
\item{...}{Objects to concatenate (each \code{dd}, \code{ddl}, or \code{fdl}).}
}
\value{
A single combined object of the same class as the first input.
}
\description{
Concatenate distributional / functional objects
}
