% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{center}
\alias{center}
\alias{center.dd}
\alias{center.ddl}
\title{Center a functional/distributional object}
\usage{
center(x, ...)

\method{center}{dd}(x, ...)

\method{center}{ddl}(x, ...)
}
\arguments{
\item{x}{A \code{dd} object or list (\code{ddl} / \code{fdl}).}

\item{...}{Further arguments forwarded to \code{\link[=dd]{dd()}}.}
}
\value{
A \code{dd} object (or list, for list input).
}
\description{
Subtract the mean (in the appropriate Bayes space) from each realization.
}
