% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cor.ij}
\alias{cor.ij}
\title{Correlation helper function}
\usage{
cor.ij(x, y, i = 1, j = 1)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{i}{a numeric value}

\item{j}{a numeric value}

\item{...}{other arguments not used by this method.}
}
\value{
Invisibly returns the correlation between two vectors raised to the power of i and j.
}
\description{
Calculates the correlation between two vectors raised to the power of i and j.
}
