% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{cov.default}
\alias{cov.default}
\title{FUNCTION_TITLE}
\usage{
\method{cov}{default}(
  x,
  y = NULL,
  use = "everything",
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
