% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{cov}
\alias{cov}
\alias{cov.default}
\alias{cov.dd}
\title{Covariance}
\usage{
cov(x, ...)

\method{cov}{default}(x, ...)

\method{cov}{dd}(x, ...)
}
\arguments{
\item{x}{A numeric vector / matrix or a \code{dd} object.}

\item{...}{Further arguments forwarded to the underlying method.}
}
\value{
A covariance estimate.
}
\description{
Generic dispatching to \code{\link[stats:cor]{stats::cov()}} by default and to \code{\link[fda:var.fd]{fda::var.fd()}}
for \code{dd} input.
}
