% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcor_test.r
\name{print.dda.nlcor}
\alias{print.dda.nlcor}
\title{Print Method for \code{dda.nlcor} Objects}
\usage{
\method{print}{dda.nlcor}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dda.nlcor}.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
\code{dda.nlcor} test statistics and p-values.
}
\description{
Displays non-linear correlation tests results between two variables.
}
