% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda_bagging.R
\name{dda_bagging}
\alias{dda_bagging}
\title{Bootstrap Aggregated DDA Analysis}
\usage{
dda_bagging(
  dda_result,
  iter = 100,
  progress = TRUE,
  save_file = NULL,
  alpha = 0.05,
  data = NULL,
  agg_stat = c("mean", "median", "trimmed", "winsorized", "midhinge", "tukey"),
  trim_prob = 0.2
)
}
\arguments{
\item{dda_result}{Output from any DDA function (dda.indep, dda.vardist, dda.resdist)}

\item{iter}{Number of bootstrap iterations (default: 100)}

\item{progress}{Whether to show progress bar (default: TRUE)}

\item{save_file}{Optional file path to save results}

\item{alpha}{Significance level for decisions (default: 0.05)}

\item{data}{Raw data frame with ALL variables (outcome, predictor, covariates)}

\item{agg_stat}{Method for aggregating test statistics and coefficients. Options: "mean", "median", "trimmed", "winsorized", "midhinge", "tukey". P-values always use harmonic mean.}

\item{trim_prob}{Proportion of observations to be trimmed or Winsorized from each end (default: 0.20).}
}
\value{
A list containing bootstrap and aggregated results
}
\description{
To Do: Add in object storage list of all the runs + specified options for
}
