% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density.default}
\alias{density.default}
\title{FUNCTION_TITLE}
\usage{
\method{density}{default}(
  sample = NULL,
  method = c("MPL", "kernel"),
  basis = fda::create.bspline.basis(rangeval, nbasis),
  rangeval = NULL,
  nbasis = 10,
  lambda = 0,
  clr = fda::fd(sample, basisobj = basis),
  constant = NULL,
  ...
)
}
\arguments{
\item{sample}{PARAM_DESCRIPTION, Default: NULL}

\item{method}{PARAM_DESCRIPTION, Default: c("MPL", "kernel")}

\item{basis}{PARAM_DESCRIPTION, Default: fda::create.bspline.basis(rangeval, nbasis)}

\item{rangeval}{PARAM_DESCRIPTION, Default: NULL}

\item{nbasis}{PARAM_DESCRIPTION, Default: 10}

\item{lambda}{PARAM_DESCRIPTION, Default: 0}

\item{clr}{PARAM_DESCRIPTION, Default: fda::fd(sample, basisobj = basis)}

\item{constant}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[fda]{create.bspline.basis}}, \code{\link[fda]{fd}}, \code{\link[fda]{fdPar}}, \code{\link[fda]{density.fd}}, \code{\link[fda]{eval.fd}}
}
