% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.fd.R
\name{density}
\alias{density}
\alias{density.fd}
\alias{loglfnden}
\alias{Varfnden}
\alias{normden.phi}
\alias{expectden.phi}
\alias{expectden.phiphit}
\alias{polintarray}
\title{FUNCTION_TITLE}
\usage{
density(x, ...)

\method{density}{fd}(
  x,
  WfdParobj,
  conv = 1e-04,
  iterlim = 20,
  active = 1:nbasis,
  dbglev = 0,
  ...
)

loglfnden(x, f, basisobj, cvec = FALSE)

Varfnden(x, basisobj, cvec = FALSE)

normden.phi(basisobj, cvec, JMAX = 15, EPS = 1e-07)

expectden.phi(basisobj, cvec, Cval = 1, nderiv = 0, JMAX = 15, EPS = 1e-07)

expectden.phiphit(
  basisobj,
  cvec,
  Cval = 1,
  nderiv1 = 0,
  nderiv2 = 0,
  JMAX = 15,
  EPS = 1e-07
)

polintarray(xa, ya, x0)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{WfdParobj}{PARAM_DESCRIPTION}

\item{conv}{PARAM_DESCRIPTION, Default: 1e-04}

\item{iterlim}{PARAM_DESCRIPTION, Default: 20}

\item{active}{PARAM_DESCRIPTION, Default: 1:nbasis}

\item{dbglev}{PARAM_DESCRIPTION, Default: 0}

\item{f}{PARAM_DESCRIPTION}

\item{basisobj}{PARAM_DESCRIPTION}

\item{cvec}{PARAM_DESCRIPTION}

\item{JMAX}{PARAM_DESCRIPTION, Default: 15}

\item{EPS}{PARAM_DESCRIPTION, Default: 1e-07}

\item{Cval}{PARAM_DESCRIPTION, Default: 1}

\item{nderiv}{PARAM_DESCRIPTION, Default: 0}

\item{nderiv1}{PARAM_DESCRIPTION, Default: 0}

\item{nderiv2}{PARAM_DESCRIPTION, Default: 0}

\item{xa}{PARAM_DESCRIPTION}

\item{ya}{PARAM_DESCRIPTION}

\item{x0}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
