% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density_kernel}
\alias{density_kernel}
\title{FUNCTION_TITLE}
\usage{
density_kernel(
  x,
  bw = "nrd0",
  adjust = 1,
  kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight",
    "cosine", "optcosine"),
  weights = NULL,
  window = kernel,
  width,
  give.Rkern = FALSE,
  subdensity = FALSE,
  warnWbw = var(weights) > 0,
  n = 512,
  from,
  to,
  cut = 3,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{sample}{PARAM_DESCRIPTION, Default: NULL}

\item{method}{PARAM_DESCRIPTION, Default: c("MPL", "kernel")}

\item{basis}{PARAM_DESCRIPTION, Default: fda::create.bspline.basis(rangeval, nbasis)}

\item{rangeval}{PARAM_DESCRIPTION, Default: NULL}

\item{nbasis}{PARAM_DESCRIPTION, Default: 10}

\item{lambda}{PARAM_DESCRIPTION, Default: 0}

\item{clr}{PARAM_DESCRIPTION, Default: fda::fd(sample, basisobj = basis)}

\item{constant}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[fda]{create.bspline.basis}}, \code{\link[fda]{fd}}, \code{\link[fda]{fdPar}}, \code{\link[fda]{density.fd}}, \code{\link[fda]{eval.fd}}
}
