% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval.dd}
\alias{eval.dd}
\title{FUNCTION_TITLE}
\usage{
\method{eval}{dd}(ddobj, t)
}
\arguments{
\item{ddobj}{PARAM_DESCRIPTION}

\item{t}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[fda]{eval.fd}}
}
