% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval.default}
\alias{eval.default}
\title{FUNCTION_TITLE}
\usage{
\method{eval}{default}(
  expr,
  envir = parent.frame(),
  enclos = if (is.list(envir) || is.pairlist(envir)) parent.frame() else baseenv()
)
}
\arguments{
\item{expr}{PARAM_DESCRIPTION}

\item{envir}{PARAM_DESCRIPTION, Default: parent.frame()}

\item{enclos}{PARAM_DESCRIPTION, Default: if (is.list(envir) || is.pairlist(envir)) parent.frame() else baseenv()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
