% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gmean.dd}
\alias{gmean.dd}
\title{FUNCTION_TITLE}
\usage{
gmean.dd(ddobj, ...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[fda]{mean.fd}}
}
