% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gmean}
\alias{gmean}
\alias{gmean.ddl}
\alias{gmean.dd}
\title{Geometric mean in Bayes space}
\usage{
gmean(x, ...)

\method{gmean}{ddl}(x, ...)

\method{gmean}{dd}(x, ...)
}
\arguments{
\item{x}{A \code{dd} object or a \code{ddl} list.}

\item{...}{Further arguments forwarded to the underlying \code{\link[=mean.dd]{mean.dd()}}.}
}
\value{
A \code{dd} object (or list for \code{ddl} input).
}
\description{
Bayes-space mean of a \code{dd} object or a list (\code{ddl}) of them.
}
