% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS.R
\name{ICS.dd}
\alias{ICS.dd}
\title{FUNCTION_TITLE}
\usage{
\method{ICS}{dd}(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
