% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS.R
\name{ICS.fd}
\alias{ICS.fd}
\title{FUNCTION_TITLE}
\usage{
\method{ICS}{fd}(fdobj, ...)
}
\arguments{
\item{fdobj}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[fda]{inprod}}, \code{\link[fda]{fd}}
\code{\link[ICS]{ICS-S3}}, \code{\link[ICS]{ics}}, \code{\link[ICS]{ICS}}
}
