% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{inprod.dd}
\alias{inprod.dd}
\title{Inner product}
\usage{
\method{inprod}{dd}(dd1, ...)
}
\arguments{
\item{...}{The other density to compute inner product.}
}
\value{
The inner product.
}
\description{
Computes the Aitchison inner product.
}
