% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{max.entropy}
\alias{max.entropy}
\title{Entropy helper function}
\usage{
\method{max}{entropy}(x)
}
\arguments{
\item{x}{a vector}

\item{...}{other arguments not used by this method.}
}
\value{
Invisibly returns the entropy of a vector
}
\description{
Calculates the entropy of a vector.
}
