% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{mean.dd}
\alias{mean.dd}
\title{Aitchison mean}
\usage{
\method{mean}{dd}(...)
}
\arguments{
\item{...}{The vector of densities.}
}
\value{
A density object which is the mean.
}
\description{
Computes the mean of a vector of densities.
}
