% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{mean.dd}
\alias{mean.dd}
\alias{mean.ddl}
\title{Mean of distributional data}
\usage{
\method{mean}{dd}(x, ...)

\method{mean}{ddl}(x, ...)
}
\arguments{
\item{x}{A \code{dd} object (or list of them) / a \code{ddl} list.}

\item{...}{Further arguments forwarded to \code{\link[=dd]{dd()}}.}
}
\value{
A \code{dd} object.
}
\description{
Mean of distributional data
}
