% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{mean_dd}
\alias{mean_dd}
\title{Aitchison mean}
\usage{
mean_dd(...)
}
\arguments{
\item{...}{The vector of densities.}
}
\value{
A density object which is the mean.
}
\description{
Computes the mean of a vector of densities.
}
