% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mysd}
\alias{mysd}
\title{Standard Deviation helper function}
\usage{
mysd(x)
}
\arguments{
\item{x}{a vector}

\item{...}{other arguments not used by this method.}
}
\value{
Invisibly returns the standard deviation of a vector
}
\description{
Calculates the standard deviation of a vector.
}
