% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{nlcorTest}
\alias{nlcorTest}
\title{Non-linear correlation (\code{nlcor}) tests}
\usage{
nlcorTest(x, y, fun, fname = NULL)
}
\arguments{
\item{x}{a numeric vector representing the tentative predictor.}

\item{y}{a numeric vector representing the tentative outcome.}

\item{fun}{a numeric value or a function of .Primitive type used for non-linear correlation tests. When \code{fun} is numeric the value is used in a power transformation.}

\item{...}{other arguments not used by this method.}
}
\value{
Invisibly returns the correlation coefficient, t-value, degrees of freedom, and p-value for three different tests.
}
\description{
Calculates the non-linear correlation between two vectors.
}
