% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cdda.indep.R
\name{plot.cddaindep}
\alias{plot.cddaindep}
\title{Plot of \code{cdda.indep} class object}
\usage{
\method{plot}{cddaindep}(obj = NULL, stat = NULL, ylim = NULL)
}
\arguments{
\item{obj}{a \code{cddaindep} class object}

\item{stat}{a character indicating the stat of statistic to be plotted with options \code{c("hsic.diff", "dcor.diff", "mi.diff")}}

\item{ylim}{a numeric vector of length 2 indicating the y-axis limits if \code{NULL}, the function will set the limits automatically}
}
\value{
A plot of a test statistic result from the \code{cdda.indep} class object.
}
\description{
\code{plot} returns test statistic results from the \code{cdda.indep} class object.
}
\examples{
           car.indep <- cdda.indep(mpg ~ wt * hp + qsec, pred = "wt",
                                             mod = "hp", diff = TRUE, data = mtcars)
                     plot(car.indep, stat = "mi.diff")

}
