% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cdda.vardist.R
\name{plot.cddavardist}
\alias{plot.cddavardist}
\title{Plot of \code{cdda.vardist} class object}
\usage{
\method{plot}{cddavardist}(
  obj = NULL,
  stat = NULL,
  ylim = NULL,
  alpha = 0.05,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{obj}{a \code{cddavardist} class object}

\item{stat}{a character indicating the statistic to be plotted, default is \code{"rhs"}, with options \code{c("rhs", "cokurt", "rcc", "rtanh")}}

\item{ylim}{a numeric vector of length 2 indicating the y-axis limits if \code{NULL}, the function will set the limits automatically}
}
\value{
A plot of a test statistic result from the \code{cdda.vardist} class object.
}
\description{
\code{plot} returns test statistic results from the \code{cdda.vardist} class object.
}
\examples{
    car.var <- cdda.vardist(mpg ~ wt * hp + qsec, pred = "wt",
                                      mod = "hp", data = mtcars)
              plot(car.var, stat = "rcc")

}
