% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R, R/plot.R
\name{plot.fdl}
\alias{plot.fdl}
\alias{plot.dd}
\alias{plot.ddl}
\title{Plot a distributional-data object}
\usage{
\method{plot}{fdl}(x, ...)

\method{plot}{dd}(x, ...)

\method{plot}{ddl}(x, ...)
}
\arguments{
\item{x}{A \code{dd} object or list (\code{ddl} / \code{fdl}).}

\item{...}{Aesthetic arguments forwarded to ggplot layers.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a distributional-data object
}
