% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdda.indep.R
\name{print.cdda.indep}
\alias{print.cdda.indep}
\title{Print method for \code{cdda.indep}}
\usage{
\method{print}{cdda.indep}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cdda.indep}.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
An object of class \code{cdda.indep} with readable OLS coefficients for competing models
}
\description{
Displays the output of standard linear model coefficients for competing target and alternative models.
}
