% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdda.indep.R
\name{print.cddaindep}
\alias{print.cddaindep}
\title{Print method for "cddaindep" class}
\usage{
\method{print}{cddaindep}(x, ...)
}
\arguments{
\item{x}{An object of class `cddaindep`}
}
\value{
An object of class \code{cddaindep} with readable ols coefficients for competing models
}
\description{
Calling `print` on a `cddaindep` object will display the output of the standard linear model coefficients for competing models.
}
