% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dda_bagging.R
\name{print.dda_bagging_indep}
\alias{print.dda_bagging_indep}
\title{Print for dda_bagging Output (INDEP)}
\usage{
\method{print}{dda_bagging_indep}(object, digits = 4, alpha = 0.05)
}
\arguments{
\item{object}{Output from dda_bagging() for dda.indep objects (class: dda_bagging_indep)}

\item{digits}{Number of digits for rounding (default: 4)}

\item{alpha}{Significance level for decisions (default: 0.05)}
}
\description{
Print for dda_bagging Output (INDEP)
}
