% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dda_bagging.R
\name{print.dda_bagging_vardist}
\alias{print.dda_bagging_vardist}
\title{Print for dda_bagging Output (VARDIST)}
\usage{
\method{print}{dda_bagging_vardist}(object, digits = 4)
}
\description{
Print for dda_bagging Output (VARDIST)
}
