% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda_indep.r
\name{print.ddaindep}
\alias{print.ddaindep}
\title{Print method for \code{ddaindep} class}
\usage{
\method{print}{ddaindep}(object, ...)
}
\arguments{
\item{x}{An object of class \code{cddaindep}}
}
\value{
An object of class \code{ddaindep} with readable ols coefficients for competing models
}
\description{
Calling \code{print} on a \code{ddaindep}object will display the results of the skewness and kurtosis tests, and bootstrap confidence intervals for the difference in skewness and kurtosis of the indepence properties of two competing models.
}
