% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda_resdist.r
\name{print.ddaresdist}
\alias{print.ddaresdist}
\title{Print method for \code{ddaresdist} objects}
\usage{
\method{print}{ddaresdist}(object)
}
\description{
Calling \code{print} on a \code{ddaresdist} object will display the test statistics and p-values for the difference in skewness and kurtosis between the residuals of the two competing models. If bootstrap confidence intervals are computed, the lower and upper bounds of the confidence intervals will also be displayed.
}
