% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda_vardist.r
\name{print.ddavardist}
\alias{print.ddavardist}
\title{Print method for \code{ddavardist} objects}
\usage{
\method{print}{ddavardist}(object)
}
\description{
Calling \code{print} on a \code{ddavardist} object will display the results of the skewness and kurtosis tests, and bootstrap confidence intervals for the difference in skewness and kurtosis of the variables of two competing models.
}
