% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dda_ols_bagging.R
\name{print_ols_summary}
\alias{print_ols_summary}
\title{Print OLS Summary from Bagged DDA}
\usage{
print_ols_summary(
  object,
  agg_stat = NULL,
  trim_prob = 0.1,
  win_prob = 0.1,
  digits = 4,
  ...
)
}
\arguments{
\item{object}{Output from dda_bagging()}

\item{agg_stat}{Method for aggregating test statistics. If NULL, uses the method established in dda_bagging().}

\item{trim_prob}{Proportion of observations to be trimmed (default: 0.10).}

\item{win_prob}{Proportion of observations to be Winsorized (default: 0.10).}

\item{digits}{Number of digits for rounding}

\item{...}{Additional arguments passed to print}
}
\description{
Print OLS Summary from Bagged DDA
}
