% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{relative}
\alias{relative}
\alias{relative.dd}
\alias{relative.ddl}
\title{Bayes-space relative density}
\usage{
relative(e1, e2, ...)

\method{relative}{dd}(e1, e2, ...)

\method{relative}{ddl}(e1, e2, ...)
}
\arguments{
\item{e1, e2}{\code{dd} objects (or lists thereof) to compare.}

\item{...}{Further arguments forwarded to the underlying method.}
}
\value{
A \code{dd} object.
}
\description{
Generic \code{relative(e1, e2)} returns the Bayes-space ratio of two
densities. The default method dispatches to \code{\link[fda:arithmetic.fd]{fda::minus.fd()}} on the
clr-transformed log-densities.
}
