% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cdda.indep.R
\name{summary.cddaindep}
\alias{summary.cddaindep}
\title{Summary of \code{cdda.indep} class object}
\usage{
\method{summary}{cddaindep}(
  cdda.output,
  nlfun = FALSE,
  hetero = TRUE,
  hsic = TRUE,
  hsic.diff = FALSE,
  dcor = TRUE,
  dcor.diff = FALSE,
  mi.diff = FALSE,
  ...
)
}
\arguments{
\item{cdda.output}{output from \code{cdda.indep} object}

\item{nlfun}{logical, default is \code{FALSE}, includes non-linear correlation test}

\item{hetero}{logical, default is \code{TRUE}, includes Breusch-Pagan Homoscedasticity test}

\item{hsic}{logical, default is \code{TRUE}, includes Hilbert-Schmidt Independence Criterion (HSIC) test}

\item{hsic.diff}{logical, default is \code{FALSE}, includes HSIC differences}

\item{dcor}{logical, default is \code{TRUE}, includes Distance Correlation (dCor) test}

\item{dcor.diff}{logical, default is \code{FALSE}, includes dCor differences}

\item{mi.diff}{logical, default is \code{FALSE}, includes Mutual Information (MI) differences}
}
\value{
A summary of test statistic results from the \code{cdda.indep} class object.
}
\description{
\code{summary} returns test statistic results from the \code{cdda.indep} class object.
}
\examples{
             car.indep <- cdda.indep(mpg ~ wt * hp + qsec, pred = "wt",
                                               mod = "hp", diff = TRUE, data = mtcars)
                       summary(car.indep, hsic.diff = TRUE, mi.diff = TRUE)

}
