% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cdda.vardist.R
\name{summary.cddavardist}
\alias{summary.cddavardist}
\title{Summary of \code{cdda.vardist} class object}
\usage{
\method{summary}{cddavardist}(
  cdda.output,
  skew = TRUE,
  coskew = FALSE,
  kurt = TRUE,
  cokurt = FALSE,
  plot = FALSE,
  ...
)
}
\arguments{
\item{cdda.output}{output from \code{cdda.vardist} object}

\item{skew}{logical, default is \code{TRUE}, where skew differences and Agostino stats printed}

\item{coskew}{logical, default is \code{FALSE}, where co-skew differences and Hyvarinen-Smith stats printed}

\item{kurt}{logical, default is \code{TRUE}, where kurtosis differences and Anscombe stats printed}

\item{cokurt}{logical, default is \code{FALSE}, where co-kurtosis differences and Chen-Chan stats printed}
}
\value{
A summary of test statistic results from the \code{cdda.vardist} class object.
}
\description{
\code{summary} returns test statistic results from the \code{cdda.vardist} class object.
}
\examples{
           car.var <- cdda.vardist(mpg ~ wt * hp + qsec, pred = "wt",
                                              mod = "hp", data = mtcars)
                     summary(car.var, coskew = TRUE, cokurt = TRUE)

}
