% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dda_bagging.R
\name{summary.dda_bagging_indep}
\alias{summary.dda_bagging_indep}
\title{Summary for dda_bagging Output (INDEP)}
\usage{
\method{summary}{dda_bagging_indep}(object, show = NULL, digits = 3, ...)
}
\arguments{
\item{object}{Output from dda_bagging() for dda.indep objects}

\item{show}{Character vector of stats to show (e.g. c("hsic", "dcor", "bp", "all"))}

\item{digits}{Number of decimal places to print for proportions (default: 3)}

\item{...}{Additional arguments}
}
\description{
Summary for dda_bagging Output (INDEP)
}
