% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{var.dd}
\alias{var.dd}
\title{FUNCTION_TITLE}
\usage{
\method{var}{dd}(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[fda]{var.fd}}
}
