% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{var.default}
\alias{var.default}
\title{FUNCTION_TITLE}
\usage{
\method{var}{default}(x, y = NULL, na.rm = FALSE, use)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
