% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{var}
\alias{var}
\alias{var.default}
\alias{var.dd}
\title{Variance}
\usage{
var(x, ...)

\method{var}{default}(x, ...)

\method{var}{dd}(x, ...)
}
\arguments{
\item{x}{A numeric vector or a \code{dd} object.}

\item{...}{Further arguments forwarded to the underlying method.}
}
\value{
A variance estimate; for \code{dd} input, an object of class \code{bidd}.
}
\description{
Generic dispatching to \code{\link[stats:cor]{stats::var()}} by default and to \code{\link[fda:var.fd]{fda::var.fd()}}
for \code{dd} input.
}
