% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vietnam_provinces}
\alias{vietnam_provinces}
\title{Vietnamese provinces (geographic boundaries)}
\format{
A \code{sf} data frame with 63 rows and 4 variables:
\describe{
\item{region}{Administrative region code.}
\item{province}{Province name.}
\item{geometry}{\code{sfc} polygon geometry of the province boundary.}
\item{climate_region}{Climate-region classification.}
}
}
\usage{
vietnam_provinces
}
\description{
Administrative-province polygons with climate-region tags.
}
\keyword{datasets}
