% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vietnam_regions}
\alias{vietnam_regions}
\title{Vietnamese administrative regions}
\format{
A data frame with 6 rows and 2 variables:
\describe{
\item{code}{Region code.}
\item{name}{Region name.}
}
}
\usage{
vietnam_regions
}
\description{
Lookup table from region code to region name.
}
\keyword{datasets}
