% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{zbsplines}
\alias{zbsplines}
\title{FUNCTION_TITLE}
\usage{
zbsplines(fdobj = NULL, coefs = fdobj$coefs, basis = fdobj$basis, inv = FALSE)
}
\arguments{
\item{fdobj}{PARAM_DESCRIPTION, Default: NULL}

\item{coefs}{PARAM_DESCRIPTION, Default: fdobj$coefs}

\item{basis}{PARAM_DESCRIPTION, Default: fdobj$basis}

\item{inv}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[MASS]{ginv}}
}
