% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{as_dd}
\alias{as_dd}
\alias{as_dd.list}
\alias{as_dd.xts}
\alias{as_dd.dd}
\alias{as_dd.fd}
\alias{as_dd.numeric}
\title{FUNCTION_TITLE}
\usage{
as_dd(...)

\method{as_dd}{list}(l, full_sample = unlist(l), mc.cores = NULL, ...)

\method{as_dd}{xts}(sample, ...)

\method{as_dd}{dd}(ddobj, ...)

\method{as_dd}{fd}(fdobj, ...)

\method{as_dd}{numeric}(sample, ...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{l}{PARAM_DESCRIPTION}

\item{full_sample}{PARAM_DESCRIPTION, Default: unlist(l)}

\item{mc.cores}{PARAM_DESCRIPTION, Default: NULL}

\item{sample}{PARAM_DESCRIPTION}

\item{ddobj}{PARAM_DESCRIPTION}

\item{fdobj}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[parallel]{mclapply}}
}
