% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R, R/fd.R
\name{c.dd}
\alias{c.dd}
\alias{c.ddl}
\alias{c.fdl}
\title{FUNCTION_TITLE}
\usage{
\method{c}{dd}(...)

\method{c}{ddl}(...)

\method{c}{fdl}(l)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{l}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[fda]{character(0)}}
}
