% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdda.indep.R
\name{cdda.indep}
\alias{cdda.indep}
\title{Conditional Direction Dependence Analysis: Independence Assumptions}
\usage{
cdda.indep(
  formula = NULL,
  pred = NULL,
  mod = NULL,
  modval = "mean",
  data = list(),
  hetero = TRUE,
  diff = FALSE,
  nlfun = NULL,
  hsic.method = "gamma",
  B = 200,
  boot.type = "perc",
  conf.level = 0.95,
  parallelize = FALSE,
  cores = 1,
  values = NULL,
  JN.length = 3,
  ...
)
}
\arguments{
\item{formula}{symbolic formula of the model to be tested or a \code{lm} object.}

\item{pred}{a character indicating the variable name of the predictor which serves as the outcome in the alternative model.}

\item{mod}{a character indicating the variable name of the moderator which serves as an interaction term in models.}

\item{modval}{For continuous moderators, character or numeric sequence specifying how the moderator is treated in the model. Possible characters include c("mean", "median", "JN"), and if a numeric sequence is specified, the 'pick a point' approach will be utilized.}

\item{data}{a required data frame containing the variables in the model (by default variables are taken from the environment which dda.vardist is called from)}

\item{hetero}{a logical value indicating whether separate homoscedasticity tests (i.e., standard and robust Breusch-Pagan tests) should be computed.}

\item{diff}{A logical value indicating whether differences in HSIC, dCor, and MI values should be computed. Bootstrap confidence intervals are computed using B bootstrap samples.}

\item{nlfun}{Either a numeric value or a function of .Primitive type used for non-linear correlation tests. When nlfun is numeric the value is used in a power tranformation.}

\item{hsic.method}{a character indicating the inference method for Hilbert-Schmidt Independence Criterion. Must be one of the two values c("gamma", "boot").}

\item{B}{number of bootstrap samples. Only used if hsic.method = "boot"}

\item{parallelize}{A logical value indicating whether bootstrapping is performed on multiple cores. Only used if hsic.method = "boot".}

\item{cores}{a numeric value indicating the number of cores. Only used if parallelize = TRUE}
}
\value{
A list of class \code{cddaindep} containing with information on competing models under moderation
}
\description{
\code{cdda.indep} tests the skewness and kurtosis of the variables of two competing models. It also tests the difference in skewness and kurtosis to assess independence properties of the two models. The function also provides bootstrap confidence intervals for the difference in skewness and kurtosis.
}
\examples{
cdda.car.indep <- cdda.indep(mpg ~ wt * hp + qsec, pred = "wt",
                                   mod = "hp", diff = TRUE, data = mtcars)
          cdda.car.indep
          #OR
          car.test.mod <- lm(mpg ~ wt * hp + qsec, data = mtcars)
          cdda.indep(car.test.mod, pred = "wt", mod = "hp", diff = TRUE, data = mtcars)

}
