% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdda.indep.R
\name{cdda.indep}
\alias{cdda.indep}
\title{Conditional Direction Dependence Analysis: Independence Assumptions}
\usage{
cdda.indep(
  formula = NULL,
  pred = NULL,
  mod = NULL,
  modval = "mean",
  data = list(),
  hetero = TRUE,
  diff = FALSE,
  nlfun = NULL,
  hsic.method = "gamma",
  B = 200,
  boot.type = "perc",
  conf.level = 0.95,
  parallelize = FALSE,
  cores = 1,
  values = NULL,
  JN.length = 3,
  ...
)
}
\arguments{
\item{formula}{symbolic formula of the model to be tested or a \code{lm} object.}

\item{pred}{a character indicating the variable name of the predictor which serves as the outcome in the alternative model.}

\item{mod}{a character indicating the variable name of the moderator which serves as an interaction term in models.}

\item{modval}{For continuous moderators, character or numeric sequence specifying how the moderator is treated in the model. Possible characters include c("mean", "median", "JN"), and if a numeric sequence is specified, the 'pick a point' approach will be utilized.}

\item{data}{a required data frame containing the variables in the model (by default variables are taken from the environment which dda.vardist is called from)}

\item{hetero}{a logical value indicating whether separate homoscedasticity tests (i.e., standard and robust Breusch-Pagan tests) should be computed.}

\item{diff}{A logical value indicating whether differences in HSIC, dCor, and MI values should be computed. Bootstrap confidence intervals are computed using B bootstrap samples.}

\item{nlfun}{Either a numeric value or a function of .Primitive type used for non-linear correlation tests. When nlfun is numeric the value is used in a power tranformation.}

\item{hsic.method}{a character indicating the inference method for Hilbert-Schmidt Independence Criterion. Must be one of the two values c("gamma", "boot").}

\item{B}{number of bootstrap samples. Only used if hsic.method = "boot"}

\item{parallelize}{A logical value indicating whether bootstrapping is performed on multiple cores. Only used if hsic.method = "boot".}

\item{cores}{a numeric value indicating the number of cores. Only used if parallelize = TRUE}
}
\value{
A list of class \code{cddaindep} containing with information on competing models under moderation
}
\description{
\code{cdda.indep} computes and returns statistics that target
asymmetry in the independence components (predictor-error) competing models
\code{y ~ x | m} and \code{x ~ y | m} where \code{m} is a moderator. The function
is designed to handle different moderator cases, both categorical and continuous.
}
\examples{
n <- 1000

          ## --- generate moderator
          z <- sort(rnorm(n))
          z1 <- z[z <= 0]
          z2 <- z[z > 0]

          x1 <- rchisq(length(z1), df = 4) - 4
          e1 <- rchisq(length(z1), df = 3) - 3
          y1 <- 0.5 * x1 + e1

          ## --- y -> x when m > 0
          y2 <- rchisq(length(z2), df = 4) - 4
          e2 <- rchisq(length(z2), df = 3) - 3
          x2 <- 0.25 * y2 + e2

          y <- c(y1, y2); x <- c(x1, x2)

          dat <- data.frame(x, y, z)

          #cdda.indep(y ~ x * z, pred = "x", mod = "z", B = 500,
          #           diff = TRUE, nlfun = 2, data = dat)
          ## OR
          m <- lm(y ~ x * z, data = dat)
          test.cdda.indep <- cdda.indep(m, pred = "x", mod = "z", B = 500,
                                       diff = TRUE, nlfun = 2, data = dat)
          print(test.cdda.indep)

}
\seealso{
\code{\link{dda.indep}} for a non-conditional version of the function.
}
