% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdda.vardist.R
\name{cdda.vardist}
\alias{cdda.vardist}
\title{Conditional Directional Dependence Analysis: Variable Distribution}
\usage{
cdda.vardist(
  formula = NULL,
  pred = NULL,
  mod = NULL,
  modval = "mean",
  data = list(),
  B = 200,
  boot.type = "perc",
  conf.level = 0.95,
  JN.length = 3,
  ...
)
}
\arguments{
\item{formula}{Symbolic formula of the model to be tested or a \code{lm} object}

\item{pred}{A character indicating the variable name of the predictor which serves as the outcome in the alternative model.}

\item{mod}{A character indicating the variable name of the moderator which serves as an interaction term in models.}

\item{modval}{Characters or a numeric sequence specifying the moderator values used in post-hoc probing. Possible characters include \code{c("mean", "median", "JN")}. \code{modval = "mean"} tests the interaction effect at the moderator values M – 1SD, M, and M + 1SD; \code{modval = "median"} uses Q1, Md, and Q3. The Johnson-Neyman approach is applied when \code{modval = "JN"}. When a numeric sequence is specified, the pick-a-point approach will be utilized for the selected numeric values.}

\item{data}{A required data frame containing the variables in the model.}

\item{B}{Number of bootstrap samples.}

\item{boot.type}{A character indicating the type of bootstrap confidence intervals. Must be one of the two values \code{c("perc", "bca")}. \code{boot.type = "bca"} is the default.}

\item{conf.level}{Confidence level for boostrap confidence intervals.}

\item{JN.length}{A numeric value indicating the number of Johnson-Neyman points to be computed. Only used if \code{modval = "JN"}.}
}
\value{
A list of class \code{ddavardist} containing the results of CDDA
         tests to evaluate distributional properties of observed variables
         for pre-specified moderator values.
}
\description{
\code{cdda.vardist} computes DDA test statistics for observed
             variable distributions of competing conditional models (
             \code{y ~ x | m} vs. \code{x ~ y | m} with \code{m} being a
             continuous or categorical moderator).
}
\examples{
set.seed(123)
n <- 1000

## --- generate moderator
z <- sort(rnorm(n))
z1 <- z[z <= 0]
z2 <- z[z > 0]

x1 <- rchisq(length(z1), df = 4) - 4
e1 <- rchisq(length(z1), df = 3) - 3
y1 <- 0.5 * x1 + e1

## --- y -> x when m > 0
y2 <- rchisq(length(z2), df = 4) - 4
e2 <- rchisq(length(z2), df = 3) - 3
x2 <- 0.25 * y2 + e2

y <- c(y1, y2); x <- c(x1, x2)

d <- data.frame(x, y, z)

cdda.vardist(y ~ x * z, pred = "x", mod = "z",
             B = 500, modval = "JN", data = d)
## OR

m <- lm(y ~ x * z, data = d)
results <- cdda.vardist(m, pred = "x",
                        mod = "z", data = d, B = 500, modval = "JN")
print(results)

}
\references{
Wiedermann, W., & von Eye, A. (2025). Direction Dependence Analysis: Foundations and Statistical Methods. Cambridge, UK: Cambridge University Press.
}
\seealso{
\code{\link{dda.vardist}} for a non-conditional version of the function.
}
