% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdda.vardist.R
\name{cdda.vardist}
\alias{cdda.vardist}
\title{Conditional Directional Dependence Analysis: Variable Distribution}
\usage{
cdda.vardist(
  formula = NULL,
  pred = NULL,
  mod = NULL,
  modval = "mean",
  data = list(),
  B = 200,
  boot.type = "perc",
  conf.level = 0.95,
  values = NULL,
  JN.length = 3,
  ...
)
}
\arguments{
\item{formula}{symbolic formula of the model to be tested or a \code{lm} object}

\item{pred}{variable name of the predictor which serves as the outcome in the alternative model}

\item{mod}{a character indicating the variable name of the moderator which serves as an
interaction term in models.}

\item{modval}{For continuous moderators, character or numeric sequence specifying how the moderator is treated in the model. Possible characters include \code{c("mean", "median", "JN")}, and if a numeric sequence is specified, the 'pick a point' approach will be utilized.}

\item{data}{a required data frame containing the variables in the model (by default variables are taken from the environment which \code{dda.vardist} is called from)}

\item{B}{number of bootstrap samples}

\item{boot.type}{A vector of character strings representing the type of bootstrap confidence intervals required. Must be one of the two values \code{c("perc", "bca")}. \code{boot.type = "bca"} is the default.}

\item{conf.level}{confidence level for boostrap confidence intervals}
}
\value{
A list of class \code{cddavardist} containing the results of the conditional directional dependence analysis.
}
\description{
\code{cdda.vardist} provides information on competing models under moderation
}
\examples{
         car.indep <- cdda.vardist(mpg ~ wt * hp + qsec, pred = "wt",
                                             mod = "hp", data = mtcars) or
                   car.test <- lm(mpg ~ wt * hp + qsec, data = mtcars)
                   cdda.vardist(car.test, pred = "wt", mod = "hp", data = mtcars)

}
