% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{center}
\alias{center}
\alias{center.dd}
\alias{center.ddl}
\title{FUNCTION_TITLE}
\usage{
center(...)

\method{center}{dd}(ddobj, ...)

\method{center}{ddl}(ddlist, ...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{ddobj}{PARAM_DESCRIPTION}

\item{ddlist}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fda]{center.fd}}
}
