% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{cov.default}
\alias{cov.default}
\alias{cov}
\alias{cov.dd}
\title{FUNCTION_TITLE}
\usage{
\method{cov}{default}(
  x,
  y = NULL,
  use = "everything",
  method = c("pearson", "kendall", "spearman")
)

cov(...)

\method{cov}{dd}(...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION, Default: NULL}

\item{use}{PARAM_DESCRIPTION, Default: 'everything'}

\item{method}{PARAM_DESCRIPTION, Default: c("pearson", "kendall", "spearman")}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fda]{var.fd}}
}
