% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{dd}
\alias{dd}
\title{R6 Class Representing Distributional Data}
\description{
Using Bayes Spaces.
}
\details{
More details here.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{clr}}{An \code{fd} object containing representation of clr on a
ZB-spline basis.}

\item{\code{constant}}{Normalizing constant.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dd-new}{\code{dd$new()}}
\item \href{#method-dd-+}{\code{dd$+()}}
\item \href{#method-dd--}{\code{dd$-()}}
\item \href{#method-dd-*}{\code{dd$*()}}
\item \href{#method-dd-eval}{\code{dd$eval()}}
\item \href{#method-dd-normalize}{\code{dd$normalize()}}
\item \href{#method-dd-plot}{\code{dd$plot()}}
\item \href{#method-dd-inprod}{\code{dd$inprod()}}
\item \href{#method-dd-clone}{\code{dd$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dd-new"></a>}}
\if{latex}{\out{\hypertarget{method-dd-new}{}}}
\subsection{Method \code{new()}}{
Create a new distributional data object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dd$new(
  sample = NULL,
  basis = fda::create.bspline.basis(rangeval, nbasis),
  rangeval = NULL,
  nbasis = 10,
  lambda = 0,
  clr = fda::fd(sample, basisobj = basis),
  constant = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{An \code{fd} object representating the clr transform.}

\item{\code{basis}}{The bspline basis.}

\item{\code{rangeval}}{An \code{fd} object representating the clr transform.}

\item{\code{nbasis}}{An \code{fd} object representating the clr transform.}

\item{\code{lambda}}{An \code{fd} object representating the clr transform.}

\item{\code{clr}}{An \code{fd} object representating the clr transform.}

\item{\code{constant}}{Normalizing constant.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \verb{Distributional Data} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dd-+"></a>}}
\if{latex}{\out{\hypertarget{method-dd-+}{}}}
\subsection{Method \code{+()}}{
Add two distributional data objects together.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dd$+(dd2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dd2}}{Another \code{dd} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{dd} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dd--"></a>}}
\if{latex}{\out{\hypertarget{method-dd--}{}}}
\subsection{Method \code{-()}}{
Substract two distributional data objects together.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dd$-(dd2 = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dd2}}{Another \code{dd} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{dd} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dd-*"></a>}}
\if{latex}{\out{\hypertarget{method-dd-*}{}}}
\subsection{Method \code{*()}}{
Multiply a distributional data object with a scalar.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dd$*(alpha)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{A numerical value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{dd} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dd-eval"></a>}}
\if{latex}{\out{\hypertarget{method-dd-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluates a distributional data object at a point \code{t}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dd$eval(t)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{A real number \code{t}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A real number.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dd-normalize"></a>}}
\if{latex}{\out{\hypertarget{method-dd-normalize}{}}}
\subsection{Method \code{normalize()}}{
Computes the normalizing constant of distributional data object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dd$normalize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dd-plot"></a>}}
\if{latex}{\out{\hypertarget{method-dd-plot}{}}}
\subsection{Method \code{plot()}}{
Plots a distributional data object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dd$plot(rangeval = self$clr$basis$rangeval, h = 0.01, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rangeval}}{The interval on which to plot.}

\item{\code{h}}{The step between each point.}

\item{\code{...}}{The rest.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dd-inprod"></a>}}
\if{latex}{\out{\hypertarget{method-dd-inprod}{}}}
\subsection{Method \code{inprod()}}{
Computes the Aitchison inner product.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dd$inprod(dd2 = self)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dd2}}{The other density to compute inner product.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The inner product.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dd$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
