% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{dd}
\alias{dd}
\title{FUNCTION_TITLE}
\usage{
dd(
  sample = NULL,
  full_sample = sample,
  basis = fda::create.bspline.basis(rangeval = rangeval, nbasis = nbasis, norder =
    norder, breaks = breaks),
  knots_pos = "quantiles",
  breaks = if (is.numeric(sample)) {
     if (knots_pos == "quantiles") {
        
    quantile(full_sample, seq(0, 1, length.out = nbasis - norder + 2))
     }
     else
    {
         quantile(range(full_sample), seq(0, 1, length.out = nbasis - norder + 2))

        }
 } else {
     NULL
 },
  rangeval = if (is.numeric(sample)) range(breaks) else c(0, 1),
  nbasis = 12,
  norder = 5,
  lambda = 0,
  clr = fda::fd(basisobj = basis),
  constant = NULL,
  normalize = TRUE,
  ...
)
}
\arguments{
\item{sample}{PARAM_DESCRIPTION, Default: NULL}

\item{full_sample}{PARAM_DESCRIPTION, Default: sample}

\item{basis}{PARAM_DESCRIPTION, Default: fda::create.bspline.basis(rangeval = rangeval, nbasis = nbasis,
norder = norder, breaks = breaks)}

\item{knots_pos}{PARAM_DESCRIPTION, Default: 'quantiles'}

\item{breaks}{PARAM_DESCRIPTION, Default: if (is.numeric(sample)) {
if (knots_pos == "quantiles") {
quantile(full_sample, seq(0, 1, length.out = nbasis -
norder + 2))
}
else {
quantile(range(full_sample), seq(0, 1, length.out = nbasis -
norder + 2))
}
} else {
NULL
}}

\item{rangeval}{PARAM_DESCRIPTION, Default: if (is.numeric(sample)) range(breaks) else c(0, 1)}

\item{nbasis}{PARAM_DESCRIPTION, Default: 12}

\item{norder}{PARAM_DESCRIPTION, Default: 5}

\item{lambda}{PARAM_DESCRIPTION, Default: 0}

\item{clr}{PARAM_DESCRIPTION, Default: fda::fd(basisobj = basis)}

\item{constant}{PARAM_DESCRIPTION, Default: NULL}

\item{normalize}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[fda]{create.bspline.basis}}, \code{\link[fda]{fd}}, \code{\link[fda]{fdPar}}, \code{\link[fda]{eval.fd}}
\code{\link[stats]{integrate}}
}
