% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda-package.R
\docType{package}
\name{dda-package}
\alias{dda}
\alias{dda-package}
\title{dda: Distributional Data Analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Camille Mondon \email{camille.mondon@tse-fr.eu} (\href{https://orcid.org/0009-0007-4569-990X}{ORCID})

}
\keyword{internal}
