% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda-package.R
\docType{package}
\name{dda-package}
\alias{dda}
\alias{dda-package}
\title{dda: Direction Dependence Analysis}
\description{
Functions to analyze the causal direction of dependence in linear models (Wiedermann, W., & von Eye, A., 2025). The package includes functions to compute DDA test statistics for variable distributions, residual distributions, and independence properties of predictors and residuals in competing causal models. In addition, the package contains functions to test the causal direction of dependence in conditional models (i.e., models with interaction terms) For more information see \url{https://www.ddaproject.com}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wwiedermann/dda}
}

}
\author{
\strong{Maintainer}: Wolfgang Wiedermann \email{wiedermannw@missouri.edu}

Authors:
\itemize{
  \item Megan Hirni \email{mj.hirni@outlook.com}
}

}
\keyword{internal}
