% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda_indep.r
\name{dda.indep}
\alias{dda.indep}
\title{Direction Dependence Analysis: Independence Distribution}
\usage{
dda.indep(
  formula,
  pred = NULL,
  data = list(),
  nlfun = NULL,
  hetero = FALSE,
  hsic.method = "gamma",
  diff = FALSE,
  B = 200,
  boot.type = "perc",
  conf.level = 0.95,
  parallelize = FALSE,
  cores = 1,
  ...
)
}
\arguments{
\item{formula:}{symbolic formula of the model to be tested or a \code{lm} object.}

\item{pred:}{a character indicating the variable name of the predictor which serves as the outcome in the alternative model.}

\item{data:}{an optional data frame containing the variables in the model (by default variables are taken from the environment which \code{dda.indep} is called from)}

\item{nlfun:}{Either a numeric value or a function of .Primitive type used for non-linear correlation tests. When \code{nlfun} is numeric the value is used in a power tranformation.}

\item{hetero:}{A logical value indicating whether separate homoscedasticity tests (i.e., standard and robust Breusch-Pagan tests) should be computed.}

\item{hsic.method:}{a character indicating the inference method for Hilbert-Schmidt Independence Criterion (HSIC). Must be one of the four values \code{c("gamma", "eigenvalue", "boot", "permutation")}. \code{hsic.method = "gamma"} is the default.}

\item{diff:}{A logical value indicating whether differences in HSIC, Distance Correlation (dCor), and MI values should be computed. Bootstrap confidence intervals are computed using \code{B} bootstrap samples.}

\item{B:}{Number of permutations for separate dCor tests and number of resamples if \code{hsic.method = c("boot", "permutation")} or \code{diff = TRUE}}

\item{boot.type:}{A vector of character strings representing the type of bootstrap confidence intervals required. Must be one of the two values \code{c("perc", "bca")}. \code{boot.type = "perc"} is the default.}

\item{conf.level:}{confidence level for bootstrap confidence intervals}

\item{parallelize:}{A logical value indicating whether boostrapping is performed on multiple cores. Only used if \code{diff = TRUE.}}

\item{cores:}{a numeric value indicating the number of cores. Only used if parallelize = TRUE}
}
\value{
An object of class \code{ddaindep} containing the results of the independence tests.
}
\description{
\code{dda.indep} computes and returns statistics that target
             asymmetry in the independence components (predictor-error)
             competing models \code{y ~ x} and \code{x ~ y}.
}
\examples{
n <- 1000

          x1 <- rchisq(length(n * 0.5), df = 4) - 4
          e1 <- rchisq(length(n * 0.5), df = 3) - 3
          y1 <- 0.5 * x1 + e1

          ## --- y -> x when m > 0
          y2 <- rchisq(length(n * 0.5), df = 4) - 4
          e2 <- rchisq(length(n * 0.5, df = 3) - 3
          x2 <- 0.25 * y2 + e2

          y <- c(y1, y2); x <- c(x1, x2)

          m <- lm(y ~ x)
          dda.indep(y ~ x, pred = "x", data = data.frame(x, y), nlfun = 2,
                    B = 500, hetero = TRUE, diff = TRUE)

}
