% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda_resdist.r
\name{print.dda.resdist}
\alias{print.dda.resdist}
\title{Print Method for \code{dda.resdist} Objects}
\usage{
\method{print}{dda.resdist}(object, ...)
}
\arguments{
\item{object}{An object of class \code{dda.resdist}.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
An object of class \code{dda.resdist}.
}
\description{
Displays the test statistics and p-values for the asymmetry of error
             distributions of causally competing models (\code{y ~ x} vs. \code{x ~ y}).
}
\examples{
print(x)

}
