% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda_resdist.r
\name{print.dda.resdist}
\alias{print.dda.resdist}
\title{Print Method for \code{dda.resdist} Objects}
\usage{
\method{print}{dda.resdist}(object, ...)
}
\arguments{
\item{object}{An object of class \code{dda.resdist}.}

\item{...}{Additional arguments to be passed to the function.}
}
\description{
Displays the test statistics and p-values for the difference
             in skewness and kurtosis between the residuals of the two
             competing models. If bootstrap confidence intervals are computed,
             the lower and upper bounds of the confidence intervals will also be displayed.
}
