% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda_resdist.r
\name{dda.resdist}
\alias{dda.resdist}
\title{Direction Dependence Analysis: Residual Distributions}
\arguments{
\item{formula}{symbolic formula of the target model to be tested or a \code{lm} object}

\item{pred}{variable name of the predictor which serves as the outcome in the alternative model}

\item{data}{an optional data frame containing the variables in the model (by default variables are taken from the environment which \code{dda.vardist} is called from)}

\item{B}{number of bootstrap samples}

\item{boot.type}{A vector of character strings representing the type of bootstrap confidence intervals required. Must be one of the two values \code{c("perc", "bca")}; \code{boot.type = "bca" }is the default.}

\item{conf.level}{confidence level for boostrap confidence intervals}
}
\value{
An object of class \code{dda.Res} containing the results of skewness and kurtosis tests, the difference in skewness and kurtosis, and bootstrap confidence intervals for the difference in skewness and kurtosis.
}
\description{
This function tests the skewness and kurtosis of the residuals of two competing models. It also tests the difference in skewness and kurtosis between the residuals of the two models. The function also provides bootstrap confidence intervals for the difference in skewness and kurtosis.
}
\examples{
        dda.resdist(y ~ x + z, pred = "x", data = my.data) or 
                  m <- y ~ x + z 
                  dda.vardist(m, pred = "x", data = my.data)
                  
}
