% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda_vardist.r
\docType{class}
\name{dda.vardist}
\alias{dda.vardist}
\title{Direction Dependence Analysis: Variable Distributions}
\arguments{
\item{formula}{symbolic formula of the model to be tested or a \code{lm} object}

\item{pred}{variable name of the predictor which serves as the outcome in the alternative model}

\item{data}{an optional data frame containing the variables in the model (by default variables are taken from the environment which \code{dda.vardist} is called from)}

\item{conf.level}{confidence level for boostrap confidence intervals}

\item{B}{number of bootstrap samples}

\item{boot.type}{A vector of character strings representing the type of bootstrap confidence intervals required. Must be one of the two values \code{c("perc", "bca")}. \code{boot.type = "bca"} is the default.}
}
\value{
An object of class \code{dda.Var} containing the results of skewness and kurtosis tests, and bootstrap confidence intervals for the difference in skewness and kurtosis.
}
\description{
\code{dda.vardist} tests the skewness and kurtosis of the variables of two competing models. It also tests the difference in skewness and kurtosis between the residuals of the two models. The function also provides bootstrap confidence intervals for the difference in skewness and kurtosis.
}
\examples{
         dda.car.vardist <- dda.vardist(mpg ~ wt + qsec, pred = "wt",
                                                  B = 500, data = mtcars)
                   dda.car.vardist
                   # OR
                   car.test <- lm(mpg ~ wt + qsec, data = mtcars)
                   dda.vardist(car.test, pred = "wt", B = 500, data = mtcars)

}
