% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda_vardist.r
\name{dda.vardist}
\alias{dda.vardist}
\title{Direction Dependence Analysis: Variable Distributions}
\usage{
dda.vardist(
  formula,
  pred = NULL,
  data = list(),
  B = 200,
  boot.type = "perc",
  conf.level = 0.95
)
}
\arguments{
\item{formula}{symbolic formula of the model to be tested or a \code{lm} object}

\item{pred}{variable name of the predictor which serves as the outcome in the alternative model}

\item{data}{an optional data frame containing the variables in the model (by default variables are taken from the environment which \code{dda.vardist} is called from)}

\item{B}{number of bootstrap samples}

\item{boot.type}{A vector of character strings representing the type of bootstrap confidence intervals required. Must be one of the two values \code{c("perc", "bca")}. \code{boot.type = "bca"} is the default.}

\item{conf.level}{confidence level for boostrap confidence intervals}
}
\value{
An object of class \code{ddavardist} containing the results of
          of DDA tests of asymmetry patterns of variable distributions.
}
\description{
\code{dda.vardist} evaluates patterns of asymmetry of variable
             distributions for causally competing models
             (\code{y ~ x} vs. \code{x ~ y}).
}
\examples{
set.seed(123)
n <- 500

x <- rchisq(n, df = 4) - 4
e <- rchisq(n, df = 3) - 3
y <- 0.5 * x + e
d <- data.frame(x, y)
dda.vardist(y ~ x, pred = "x", data = d,
           B = 2000, boot.type = "bca")

}
\references{
Wiedermann, W., & von Eye, A. (2025). Direction Dependence Analysis: Foundations and Statistical Methods. Cambridge, UK: Cambridge University Press.
}
\seealso{
\code{\link{dda.vardist}} for a conditional version of the function.
}
