% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dda_vardist.r
\name{dda.vardist}
\alias{dda.vardist}
\title{Direction Dependence Analysis: Variable Distributions}
\usage{
dda.vardist(
  formula,
  pred = NULL,
  data = list(),
  B = 100,
  boot.type = "bca",
  conf.level = 0.95
)
}
\arguments{
\item{formula}{symbolic formula of the model to be tested or a \code{lm} object}

\item{pred}{variable name of the predictor which serves as the outcome in the alternative model}

\item{data}{an optional data frame containing the variables in the model (by default variables are taken from the environment which \code{dda.vardist} is called from)}

\item{B}{number of bootstrap samples}

\item{boot.type}{A vector of character strings representing the type of bootstrap confidence intervals required. Must be one of the two values \code{c("perc", "bca")}. \code{boot.type = "bca"} is the default.}

\item{conf.level}{confidence level for boostrap confidence intervals}
}
\value{
An object of class \code{ddavardist} containing the results of skewness and kurtosis tests, and bootstrap confidence intervals for the difference in skewness and kurtosis.
}
\description{
\code{dda.vardist} tests observed variable distribution properties
             including skewness and kurtosis properties in competing models
             \code{y ~ x} and \code{x ~ y}.
}
\examples{
n <- 1000

          x1 <- rchisq(length(n * 0.5), df = 4) - 4
          e1 <- rchisq(length(n * 0.5), df = 3) - 3
          y1 <- 0.5 * x1 + e1

          ## --- y -> x when m > 0
          y2 <- rchisq(length(n * 0.5), df = 4) - 4
          e2 <- rchisq(length(n * 0.5), df = 3) - 3
          x2 <- 0.25 * y2 + e2

          y <- c(y1, y2); x <- c(x1, x2)

          m <- lm(y ~ x)
          dda.vardist(y ~ x, pred = "x", data = data.frame(x, y), B = 500, boot.type = "perc")

}
\seealso{
\code{\link{dda.vardist}} for a conditional version of the function.
}
