% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{eval.dd}
\alias{eval.dd}
\title{Evaluation}
\usage{
\method{eval}{dd}(dd, t)
}
\arguments{
\item{dd}{Density to eval.}

\item{t}{A real number \code{t}.}
}
\value{
A real number.
}
\description{
Evaluates a distributional data object at a point \code{t}.
}
