% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval.default}
\alias{eval.default}
\alias{eval.dd}
\title{FUNCTION_TITLE}
\usage{
\method{eval}{default}(
  expr,
  envir = parent.frame(),
  enclos = if (is.list(envir) || is.pairlist(envir)) parent.frame() else baseenv()
)

\method{eval}{dd}(ddobj, t)
}
\arguments{
\item{expr}{PARAM_DESCRIPTION}

\item{envir}{PARAM_DESCRIPTION, Default: parent.frame()}

\item{enclos}{PARAM_DESCRIPTION, Default: if (is.list(envir) || is.pairlist(envir)) parent.frame() else baseenv()}

\item{ddobj}{PARAM_DESCRIPTION}

\item{t}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[fda]{eval.fd}}
}
