% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{gmean}
\alias{gmean}
\alias{gmean.ddl}
\alias{gmean.dd}
\title{FUNCTION_TITLE}
\usage{
gmean(...)

\method{gmean}{ddl}(ddlist, ...)

\method{gmean}{dd}(ddlist, ...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{ddlist}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
