% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_hsic_test.R
\name{hsic.test}
\alias{hsic.test}
\title{Hilbert-Schmidt Independence Criterion (HSIC) Test}
\usage{
hsic.test(
  model,
  x = NULL,
  hx = 1,
  hy = 1,
  B = 1000,
  parallelize = FALSE,
  cores = 2
)
}
\arguments{
\item{model}{An "lm" object representing the model to be tested.}

\item{x}{A vector specifying the target predictor that is used to test independence.}

\item{hx}{Numeric value specifying the bandwidth parameter for "x".}

\item{hy}{Numeric value specifying the bandwidth parameter for "y".}

\item{B}{Numeric value specifying the number of resamples used to compute the HSIC p-value.}

\item{parallelize}{A logical value indicating whether boostrapping is performed on multiple cores.}

\item{cores}{A numeric value indicating the number of cores. Only used if parallelize = TRUE.}
}
\description{
This function computes the Hilbert-Schmidt Independence Criterion (HSIC) test statistic for testing independence between two variables. The HSIC test is a non-parametric test that is based on the Hilbert-Schmidt norm of the cross-covariance operator. The null hypothesis is that the two variables are independent. The function also provides a p-value based on a bootstrap procedure.
}
\examples{
   m <- lm(y ~ x + z) 
             hsic.test(m, x, B = 500, parallelize = TRUE, cores = 4)     
}
