% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS.R
\name{ICS_outlier.default}
\alias{ICS_outlier.default}
\alias{ICS_outlier}
\alias{ICS_outlier.dd}
\alias{ICS_outlier.fd}
\title{FUNCTION_TITLE}
\usage{
\method{ICS_outlier}{default}(
  X,
  S1 = ICS_cov,
  S2 = ICS_cov4,
  S1_args = list(),
  S2_args = list(),
  ICS_algorithm = c("whiten", "standard", "QR"),
  method = "norm_test",
  test = "agostino.test",
  n_eig = 10000,
  level_test = 0.05,
  adjust = TRUE,
  level_dist = 0.025,
  n_dist = 10000,
  type = "smallprop",
  n_cores = NULL,
  iseed = NULL,
  pkg = "ICSOutlier",
  q_type = 7,
  ...
)

ICS_outlier(...)

\method{ICS_outlier}{dd}(
  X,
  S1 = ICS::ICS_cov,
  S2 = ICS::ICS_cov4,
  S1_args = list(),
  S2_args = list(),
  ICS_algorithm = c("whiten", "standard", "QR"),
  method = "norm_test",
  test = "agostino.test",
  n_eig = 10000,
  level_test = 0.05,
  adjust = TRUE,
  level_dist = 0.025,
  n_dist = 10000,
  type = "smallprop",
  n_cores = NULL,
  iseed = NULL,
  pkg = "ICSOutlier",
  q_type = 7,
  ...
)

\method{ICS_outlier}{fd}(
  X,
  S1 = ICS::ICS_cov,
  S2 = ICS::ICS_cov4,
  S1_args = list(),
  S2_args = list(),
  ICS_algorithm = c("whiten", "standard", "QR"),
  method = "norm_test",
  test = "agostino.test",
  n_eig = 10000,
  level_test = 0.05,
  adjust = TRUE,
  level_dist = 0.025,
  n_dist = 10000,
  type = "smallprop",
  n_cores = NULL,
  iseed = NULL,
  pkg = "ICSOutlier",
  q_type = 7,
  ...
)
}
\arguments{
\item{X}{PARAM_DESCRIPTION}

\item{S1}{PARAM_DESCRIPTION, Default: ICS_cov}

\item{S2}{PARAM_DESCRIPTION, Default: ICS_cov4}

\item{S1_args}{PARAM_DESCRIPTION, Default: list()}

\item{S2_args}{PARAM_DESCRIPTION, Default: list()}

\item{ICS_algorithm}{PARAM_DESCRIPTION, Default: c("whiten", "standard", "QR")}

\item{method}{PARAM_DESCRIPTION, Default: 'norm_test'}

\item{test}{PARAM_DESCRIPTION, Default: 'agostino.test'}

\item{n_eig}{PARAM_DESCRIPTION, Default: 10000}

\item{level_test}{PARAM_DESCRIPTION, Default: 0.05}

\item{adjust}{PARAM_DESCRIPTION, Default: TRUE}

\item{level_dist}{PARAM_DESCRIPTION, Default: 0.025}

\item{n_dist}{PARAM_DESCRIPTION, Default: 10000}

\item{type}{PARAM_DESCRIPTION, Default: 'smallprop'}

\item{n_cores}{PARAM_DESCRIPTION, Default: NULL}

\item{iseed}{PARAM_DESCRIPTION, Default: NULL}

\item{pkg}{PARAM_DESCRIPTION, Default: 'ICSOutlier'}

\item{q_type}{PARAM_DESCRIPTION, Default: 7}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
