% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{mean.dd}
\alias{mean.dd}
\alias{mean.ddl}
\title{FUNCTION_TITLE}
\usage{
\method{mean}{dd}(ddobj, ...)

\method{mean}{ddl}(ddlist, ...)
}
\arguments{
\item{ddobj}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{ddlist}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fda]{mean.fd}}
}
