% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{+.dd}
\alias{+.dd}
\alias{-.dd}
\alias{*.dd}
\alias{[.dd}
\alias{+.ddl}
\alias{-.ddl}
\alias{*.ddl}
\title{Bayes space operations}
\usage{
\method{+}{dd}(...)

\method{-}{dd}(...)

\method{*}{dd}(...)

\method{[}{dd}(...)

\method{+}{ddl}(ddl1, ddl2, ...)

\method{-}{ddl}(ddl1, ddl2, ...)

\method{*}{ddl}(ddl1, ddl2, ...)
}
\description{
Bayes space operations
}
