% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dd}
\alias{plot.dd}
\title{FUNCTION_TITLE}
\usage{
\method{plot}{dd}(
  ddobj,
  rangeval = ddobj$basis$rangeval,
  h = 0.001,
  show.legend = FALSE,
  ...
)
}
\arguments{
\item{ddobj}{PARAM_DESCRIPTION}

\item{rangeval}{PARAM_DESCRIPTION, Default: ddobj$basis$rangeval}

\item{h}{PARAM_DESCRIPTION, Default: 0.001}

\item{show.legend}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
