% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd.R, R/plot.R
\name{plot.fdl}
\alias{plot.fdl}
\alias{plot.dd}
\alias{plot.ddl}
\title{FUNCTION_TITLE}
\usage{
\method{plot}{fdl}(fdlist, ...)

\method{plot}{dd}(ddobj, ...)

\method{plot}{ddl}(ddlist, ...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{ddobj}{PARAM_DESCRIPTION}

\item{rangeval}{PARAM_DESCRIPTION, Default: ddobj$basis$rangeval}

\item{h}{PARAM_DESCRIPTION, Default: 0.001}

\item{show.legend}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
