% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_funs}
\alias{plot_funs}
\title{FUNCTION_TITLE}
\usage{
plot_funs(
  .data,
  funs,
  ...,
  n = 401,
  rangeval = range(lapply(dplyr::pull(.data, {
     {
         funs
     }
 }),
    function(fun) fun$basis$rangeval)),
  x = seq(rangeval[1], rangeval[2], length.out = n)
)
}
\arguments{
\item{.data}{PARAM_DESCRIPTION}

\item{funs}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{n}{PARAM_DESCRIPTION, Default: 401}

\item{rangeval}{PARAM_DESCRIPTION, Default: range(lapply(dplyr::pull(.data, {
{
funs
}
}), function(fun) fun$basis$rangeval))}

\item{x}{PARAM_DESCRIPTION, Default: seq(rangeval\link{1}, rangeval\link{2}, length.out = n)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[dplyr]{pull}}
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{geom_path}}
}
