% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dda_bagging.R
\name{print.dda_bagging_indep}
\alias{print.dda_bagging_indep}
\alias{print.dda_bagging_resdist}
\alias{print.dda_bagging_vardist}
\alias{print_ols_summary}
\title{Print Methods for Bagged DDA Objects}
\usage{
\method{print}{dda_bagging_indep}(
  x,
  agg_stat = NULL,
  trim_prob = 0.1,
  win_prob = 0.1,
  digits = 4,
  alpha = 0.05,
  ...
)

\method{print}{dda_bagging_resdist}(x, agg_stat = NULL, trim_prob = 0.1, win_prob = 0.1, digits = 4, ...)

\method{print}{dda_bagging_vardist}(x, agg_stat = NULL, trim_prob = 0.1, win_prob = 0.1, digits = 4, ...)

print_ols_summary(
  object,
  agg_stat = NULL,
  trim_prob = 0.1,
  win_prob = 0.1,
  digits = 4,
  ...
)
}
\arguments{
\item{x}{An object of class \code{dda_bagging_indep}, \code{dda_bagging_resdist}, or
\code{dda_bagging_vardist}.}

\item{agg_stat}{Character. Method for aggregating test statistics.
Options: \code{"mean"}, \code{"median"}, \code{"trimmed"}, \code{"winsorized"},
\code{"midhinge"}, or \code{"tukey"}. If \code{NULL}, uses the method established
in \code{dda_bagging()}.}

\item{trim_prob}{Numeric. Proportion of observations to be trimmed (default: 0.10).}

\item{win_prob}{Numeric. Proportion of observations to be Winsorized (default: 0.10).}

\item{digits}{Integer. Number of digits for rounding (default: 4).}

\item{alpha}{Numeric. Significance level for decisions (default: 0.05). Mainly applicable
to \code{dda_bagging_indep} objects.}

\item{...}{Additional arguments passed to \code{print}.}

\item{object}{Output from \code{dda_bagging()} (used specifically in \code{print_ols_summary}).}
}
\value{
Invisibly returns the re-aggregated object.
}
\description{
\code{print} returns aggregated test statistics and summaries associated with
bagged Direction Dependence Analysis (DDA) objects. It supports independence properties
(\code{dda_bagging_indep}), residual distributions (\code{dda_bagging_resdist}), variable
distributions (\code{dda_bagging_vardist}), and OLS summaries (\code{print_ols_summary}).
}
