% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dda_bagging.R
\name{print.dda_bagging_vardist}
\alias{print.dda_bagging_vardist}
\title{Print for dda_bagging Output (VARDIST)}
\usage{
\method{print}{dda_bagging_vardist}(x, agg_stat = NULL, trim_prob = 0.2, digits = 4, ...)
}
\arguments{
\item{x}{Output from dda_bagging() for dda.vardist objects}

\item{agg_stat}{Method for aggregating test statistics. If NULL, uses the method established in dda_bagging().}

\item{trim_prob}{Proportion of observations to be trimmed or Winsorized from each end (default: 0.20).}

\item{digits}{Number of digits for rounding (default: 4)}

\item{...}{Additional arguments passed to print}
}
\description{
Print for dda_bagging Output (VARDIST)
}
