% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dda_bagging.R, R/print.dda_ols_bagging.R
\name{print_ols_summary}
\alias{print_ols_summary}
\title{Print OLS Summary from Bagged DDA}
\usage{
print_ols_summary(object, agg_stat = NULL, trim_prob = 0.2, digits = 4, ...)

print_ols_summary(object, agg_stat = NULL, trim_prob = 0.2, digits = 4, ...)
}
\arguments{
\item{object}{Output from dda_bagging()}

\item{agg_stat}{Method for aggregating test statistics. If NULL, uses the method established in dda_bagging().}

\item{trim_prob}{Proportion of observations to be trimmed or Winsorized from each end (default: 0.20).}

\item{digits}{Number of digits for rounding (default: 4)}

\item{...}{Additional arguments passed to print}
}
\description{
Print OLS Summary from Bagged DDA

Print OLS Summary from Bagged DDA
}
