% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dda_bagging.R, R/print.dda_ols_bagging.R
\name{print_ols_summary}
\alias{print_ols_summary}
\title{Print OLS Summary from Bagged DDA}
\usage{
print_ols_summary(object, digits = 4)

print_ols_summary(object, digits = 4)
}
\arguments{
\item{object}{Output from dda_bagging()}

\item{digits}{Number of digits for rounding (default: 4)}
}
\description{
Print OLS Summary from Bagged DDA

Print OLS Summary from Bagged DDA
}
