% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{relative}
\alias{relative}
\alias{relative.dd}
\alias{relative.ddl}
\title{FUNCTION_TITLE}
\usage{
relative(...)

\method{relative}{dd}(...)

\method{relative}{ddl}(ddl1, ddl2, ...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{ddl1}{PARAM_DESCRIPTION}

\item{ddl2}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fda]{arithmetic.fd}}
}
