% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dda_bagging.R
\name{summary.dda_bagging_indep}
\alias{summary.dda_bagging_indep}
\alias{summary.dda_bagging_vardist}
\alias{summary.dda_bagging_resdist}
\title{Summary Methods for Bagged DDA Objects}
\usage{
\method{summary}{dda_bagging_indep}(object, show = NULL, digits = 2, ...)

\method{summary}{dda_bagging_vardist}(object, show = NULL, moment = NULL, digits = 2, ...)

\method{summary}{dda_bagging_resdist}(object, show = NULL, moment = NULL, digits = 2, ...)
}
\arguments{
\item{object}{An object of class \code{dda_bagging_indep}, \code{dda_bagging_vardist},
or \code{dda_bagging_resdist}.}

\item{show}{Character vector specifying which statistics to display decision proportions for
\code{c("all", "bp",  "dcor", "kurt", "cokurt", "mi", "nlcor", "skew", "coskew")}. If \code{NULL},
shows default statistics based on the object type.}

\item{digits}{Integer. Number of decimal places to print for proportions (default: 2).}

\item{...}{Additional arguments passed to \code{summary}.}

\item{moment}{Numeric vector indicating which moments to include in the summary
(e.g., \code{c(3, 4)}). Applicable only to \code{dda_bagging_vardist} and \code{dda_bagging_resdist}.}
}
\value{
Invisibly returns the original object.
}
\description{
\code{summary} returns the proportion of decisions supporting the target model,
alternative model, or remaining undecided (or confounding) across bootstrap iterations for
bagged Direction Dependence Analysis (DDA) objects.
}
