% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dda_bagging.R
\name{summary.dda_bagging_vardist}
\alias{summary.dda_bagging_vardist}
\title{Summary for dda_bagging Output (VARDIST)}
\usage{
\method{summary}{dda_bagging_vardist}(
  object,
  show = NULL,
  moment = NULL,
  agg_stat = NULL,
  trim_prob = 0.2,
  digits = 3,
  ...
)
}
\arguments{
\item{object}{Output from dda_bagging() for dda.vardist objects}

\item{show}{Character vector of stats to show (e.g. c("skew", "cokurt", "all"))}

\item{moment}{Numeric vector for moments to include (3, 4)}

\item{agg_stat}{Method for aggregating test statistics. Options: "mean", "median", "trimmed", "winsorized", "midhinge", "tukey". If NULL, uses the aggregation method established in dda_bagging().}

\item{trim_prob}{Proportion of observations to be trimmed or Winsorized from each end (default: 0.20).}

\item{digits}{Number of decimal places to print for proportions (default: 3)}

\item{...}{Additional arguments}
}
\description{
Summary for dda_bagging Output (VARDIST)
}
