% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dda_ols_bagging.R
\name{summary_ols}
\alias{summary_ols}
\title{Print OLS Model Summary for Bootstrap Aggregated DDA}
\usage{
summary_ols(
  object,
  agg_stat = NULL,
  trim_prob = 0.1,
  win_prob = 0.1,
  digits = 4,
  ...
)
}
\arguments{
\item{object}{Output from \code{dda.bagging()}.}

\item{agg_stat}{Character. Specifies the method used for aggregating test
statistics and coefficients across bootstrap samples. Must be one of the
following specifications \code{c("mean", "median", "trimmed",
"winsorized", "midhinge", "tukey")}. If \code{NULL}, the function
uses the method applied with \code{dda.bagging()}.}

\item{trim_prob}{Numeric. Proportion of observations to be trimmed on each
side of the sampling distribution when \code{agg.stat = "trimmed"}
(default: 0.10).}

\item{win_prob}{Numeric. Proportion of observations to be winsorized on
each side of the sampling distribution when
\code{agg.stat = "winsorized"} (default: 0.10).}

\item{digits}{Integer. Number of digits used for rounding.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
\code{ols.summary} prints aggregated ordinary least
squares (OLS) regression summaries from a bootstrap aggregated DDA object.
Regression coefficients and standard errors are aggregated across bootstrap
samples using the method specified in \code{dda.bagging()} or overridden
via \code{agg.stat}.
}
\examples{
\donttest{
set.seed(123)
n <- 200
x <- rchisq(n, df = 4) - 4
e <- rchisq(n, df = 3) - 3
y <- 0.5 * x + e
d <- data.frame(x, y)

base_model <- dda.indep(y ~ x, pred = "x", data = d, B = 20)
bagged <- dda.bagging(base_model, data = d, iter = 10, inner_B = 20,
                       progress = FALSE)

# Print aggregated OLS coefficients for target and alternative models
summary_ols(bagged)

# Override aggregation method
summary_ols(bagged, agg_stat = "median")
}
}
