% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{var}
\alias{var}
\title{FUNCTION_TITLE}
\usage{
var(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
