% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{var.default}
\alias{var.default}
\alias{var}
\alias{var.dd}
\title{FUNCTION_TITLE}
\usage{
\method{var}{default}(x, y = NULL, na.rm = FALSE, use)

var(...)

\method{var}{dd}(...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION, Default: NULL}

\item{na.rm}{PARAM_DESCRIPTION, Default: FALSE}

\item{use}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fda]{var.fd}}
}
