% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vietnam_provinces}
\alias{vietnam_provinces}
\title{DATASET_TITLE}
\format{
A data frame with 63 rows and 2 variables:
\describe{
\item{\code{province}}{character COLUMN_DESCRIPTION}
\item{\code{region}}{character COLUMN_DESCRIPTION}
}
}
\usage{
vietnam_provinces
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
