% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vietnam_regions}
\alias{vietnam_regions}
\title{DATASET_TITLE}
\format{
A data frame with 6 rows and 2 variables:
\describe{
\item{\code{code}}{character COLUMN_DESCRIPTION}
\item{\code{name}}{character COLUMN_DESCRIPTION}
}
}
\usage{
vietnam_regions
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
