% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vietnam_temperature}
\alias{vietnam_temperature}
\title{DATASET_TITLE}
\format{
A data frame with 1890 rows and 5 variables:
\describe{
\item{\code{year}}{character COLUMN_DESCRIPTION}
\item{\code{region}}{character COLUMN_DESCRIPTION}
\item{\code{province}}{character COLUMN_DESCRIPTION}
\item{\code{t_max}}{list COLUMN_DESCRIPTION}
\item{\code{t_min}}{list COLUMN_DESCRIPTION}
}
}
\usage{
vietnam_temperature
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
